Clean some stuff up
