+++ package.json
@@ -53,6 +53,7 @@
     "react-hot-loader": "^3.0.0-beta.6",
     "react-tap-event-plugin": "^2.0.1",
     "react-virtualized": "^9.0.0",
+    "request-idle-polyfill": "1.0.1",
     "rxjs": "^5.0.2"
   },
   "devDependencies": {
